+++ el_pagination/__init__.py
@@ -5,7 +5,7 @@ Twitter-style and Digg-style pagination.
 from __future__ import unicode_literals
 
 
-VERSION = (2, 1, 2)
+VERSION = (3, 2, 4)
 
 
 def get_version():