+++ el_pagination/models.py
@@ -1,12 +1,11 @@
 """Ephemeral models used to represent a page and a list of pages."""
-
 from __future__ import unicode_literals
 
 from django.template import (
     loader,
     Context,
 )
-from django.utils.encoding import iri_to_uri
+from django.utils.encoding import iri_to_uri, force_text
 
 from el_pagination import (
     loaders,
@@ -14,12 +13,11 @@ from el_pagination import (
     utils,
 )
 
-
 # Page templates cache.
 _template_cache = {}
 
 
-class ELPage(utils.UnicodeMixin):
+class ELPage(object):
     """A page link representation.
 
     Interesting attributes:
@@ -29,6 +27,7 @@ class ELPage(utils.UnicodeMixin):
           (usually the page number as string);
         - *self.url*: the url of the page (strting with "?");
         - *self.path*: the path of the page;
+
         - *self.is_current*: return True if page is the current page displayed;
         - *self.is_first*: return True if page is the first page;
         - *self.is_last*:  return True if page is the last page.
@@ -40,52 +39,66 @@ class ELPage(utils.UnicodeMixin):
             context=None):
         self._request = request
         self.number = number
-        self.label = utils.text(number) if label is None else label
+        self.label = force_text(number) if label is None else label
         self.querystring_key = querystring_key
         self.context = context or {}
+        self.context['request'] = request
 
         self.is_current = number == current_number
         self.is_first = number == 1
         self.is_last = number == total_number
+        if settings.USE_NEXT_PREVIOUS_LINKS:
+            self.is_previous = label and number == current_number - 1
+            self.is_next = label and number == current_number + 1
 
         self.url = utils.get_querystring_for_page(
-            request, number, self.querystring_key,
-            default_number=default_number)
+            request, number,
+            self.querystring_key,
+            default_number=default_number
+        )
         path = iri_to_uri(override_path or request.path)
         self.path = '{0}{1}'.format(path, self.url)
 
-    def __unicode__(self):
+    def render_link(self):
         """Render the page as a link."""
-        context = Context({
+        extra_context = {
             'add_nofollow': settings.ADD_NOFOLLOW,
             'page': self,
             'querystring_key': self.querystring_key,
-        })
-        context.update(self.context)
+        }
         if self.is_current:
             template_name = 'el_pagination/current_link.html'
         else:
             template_name = 'el_pagination/page_link.html'
-        template = _template_cache.setdefault(
-            template_name, loader.get_template(template_name))
-        return template.render(context)
+            if settings.USE_NEXT_PREVIOUS_LINKS:
+                if self.is_previous:
+                    template_name = 'el_pagination/previous_link.html'
+                if self.is_next:
+                    template_name = 'el_pagination/next_link.html'
+        if template_name not in _template_cache:
+            _template_cache[template_name] = loader.get_template(template_name)
+        template = _template_cache[template_name]
+        with self.context.push(**extra_context):
+            return template.render(self.context.flatten())
 
 
-class PageList(utils.UnicodeMixin):
+class PageList(object):
     """A sequence of endless pages."""
 
     def __init__(
-            self, request, page, querystring_key,
-            default_number=None, override_path=None, context=None):
+            self, request, page, querystring_key, context,
+            default_number=None, override_path=None):
         self._request = request
         self._page = page
-        self.context = context or {}
+        self.context = context
+        self.context['request'] = request
         if default_number is None:
             self._default_number = 1
         else:
             self._default_number = int(default_number)
         self._querystring_key = querystring_key
         self._override_path = override_path
+        self._pages_list = []
 
     def _endless_page(self, number, label=None):
         """Factory function that returns a *ELPage* instance.
@@ -126,7 +139,7 @@ class PageList(utils.UnicodeMixin):
         for i in range(len(self)):
             yield self[i + 1]
 
-    def __unicode__(self):
+    def __str__(self):
         """Return a rendered Digg-style pagination (by default).
 
         The callable *settings.PAGE_LIST_CALLABLE* can be used to customize
@@ -151,7 +164,10 @@ class PageList(utils.UnicodeMixin):
         generating a Digg-style pagination. The value of
         *settings.PAGE_LIST_CALLABLE* can also be a dotted path to a callable.
         """
-        if len(self) > 1:
+        return ''
+
+    def get_pages_list(self):
+        if not self._pages_list:
             callable_or_path = settings.PAGE_LIST_CALLABLE
             if callable_or_path:
                 if callable(callable_or_path):
@@ -174,10 +190,14 @@ class PageList(utils.UnicodeMixin):
                     pages.append(self.last_as_arrow())
                 else:
                     pages.append(self[item])
-            context = Context({'pages': pages})
-            context.update(self.context)
-            return loader.render_to_string('el_pagination/show_pages.html',
-                                           context)
+            self._pages_list = pages
+        return self._pages_list
+
+    def get_rendered(self):
+        if len(self) > 1:
+            template = loader.get_template('el_pagination/show_pages.html')
+            with self.context.push(pages=self.get_pages_list()):
+                return template.render(self.context.flatten())
         return ''
 
     def current(self):