+++ el_pagination/settings.py
@@ -22,6 +22,9 @@ ORPHANS = getattr(settings, 'EL_PAGINATION_ORPHANS', 0)
 LOADING = getattr(
     settings, 'EL_PAGINATION_LOADING', 'loading')
 
+USE_NEXT_PREVIOUS_LINKS = getattr(
+    settings, 'EL_PAGINATION_USE_NEXT_PREVIOUS_LINKS', False)
+
 # Labels for previous and next page links.
 PREVIOUS_LABEL = getattr(
     settings, 'EL_PAGINATION_PREVIOUS_LABEL', '&lt;')
@@ -54,3 +57,7 @@ DEFAULT_CALLABLE_ARROWS = getattr(
 # Template variable name for *page_template* decorator.
 TEMPLATE_VARNAME = getattr(
     settings, 'EL_PAGINATION_TEMPLATE_VARNAME', 'template')
+
+# If page out of range, throw a 404 exception
+PAGE_OUT_OF_RANGE_404 = getattr(
+    settings, 'EL_PAGINATION_PAGE_OUT_OF_RANGE_404', False)