+++ tests/project/urls.py
@@ -2,8 +2,14 @@
 
 from __future__ import unicode_literals
 
-from django.conf.urls import patterns, url
+from django.conf import settings
 from django.views.generic import TemplateView
+from django.contrib.staticfiles.urls import staticfiles_urlpatterns
+try:
+    from django.urls import re_path as url, include
+except:
+    from django.conf.urls import url, include
+
 
 from el_pagination.decorators import (
     page_template,
@@ -14,7 +20,7 @@ from project.views import generic
 
 
 # Avoid lint errors for the following Django idiom: flake8: noqa.
-urlpatterns = patterns('',
+urlpatterns = [
     url(r'^$',
         TemplateView.as_view(template_name="home.html"),
         name='home'),
@@ -39,6 +45,10 @@ urlpatterns = patterns('',
         page_template('onscroll/page.html')(generic),
         {'template': 'onscroll/index.html'},
         name='onscroll'),
+    url(r'^feed-wrapper/$',
+        page_template('feed_wrapper/page.html')(generic),
+        {'template': 'feed_wrapper/index.html'},
+        name='feed-wrapper'),
     url(r'^chunks/$',
         page_templates({
             'chunks/objects_page.html': None,
@@ -58,4 +68,12 @@ urlpatterns = patterns('',
         page_template('callbacks/page.html')(generic),
         {'template': 'callbacks/index.html'},
         name='callbacks'),
-)
+]
+
+if settings.DEBUG:
+    if 'debug_toolbar' in settings.INSTALLED_APPS:
+        import debug_toolbar
+        urlpatterns += [url(r'^__debug__/', include(debug_toolbar.urls)), ]
+
+urlpatterns += staticfiles_urlpatterns()
+